pyproject.toml
[tool.poetry]
name = "yaman"
version = "0.1.0"
description = "Yakuake terminal manager"
authors = ["Rok Andrée <rok@andree.si>"]
license = "GPLv3"
readme = "README.md"
keywords = ["terminal", "yakuake"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Topic :: Utilities"
]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
click = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yaman = 'yaman:cli'